multi-tenant-saas-platform/infra/requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0